+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DChmNTBP.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-eu9fqBs0.js";import{_ as a}from"./mermaid.core-BAJrQQ7x.js";import"./chunk-5VM5RSS4-Coex2eSc.js";import"./chunk-XXDRQBXY-CvEu0bgh.js";import"./chunk-POPQ4Y6H-Bg58cG2u.js";import"./chunk-F27PBJKO-Br1AWatK.js";import"./channel-Bv3Vbr5Z.js";import"./index-DVWdeaVw.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};